+++ src/Database/PostgreSQL/Simple/ToRow.hs
@@ -21,6 +21,8 @@ module Database.PostgreSQL.Simple.ToRow
       ToRow(..)
     ) where
 
+import Data.Foldable (toList)
+import Data.List.NonEmpty (NonEmpty)
 import Database.PostgreSQL.Simple.ToField (Action(..), ToField(..))
 import Database.PostgreSQL.Simple.Types (Only(..), (:.)(..))
 import GHC.Generics
@@ -197,6 +199,9 @@ instance (ToField a, ToField b, ToField c, ToField d, ToField e, ToField f,
 
 instance (ToField a) => ToRow [a] where
     toRow = map toField
+
+instance (ToField a) => ToRow (NonEmpty a) where
+    toRow = map toField . toList
 
 instance (ToRow a, ToRow b) => ToRow (a :. b) where
     toRow (a :. b) = toRow a ++ toRow b